Changes done in 24.4.01 on Grafana 11.4.2

+++ pkg/api/datasources.go
@@ -11,6 +11,7 @@ import (
 	"strconv"
 	"strings"
 
+	kp "github.com/grafana/grafana/pkg/bmc/kafkaproducer"
 	"github.com/prometheus/prometheus/promql/parser"
 	"golang.org/x/exp/slices"
 
@@ -48,6 +49,12 @@ var secretsPluginError datasources.ErrDatasourceSecretsPluginUserFriendly
 // 500: internalServerError
 func (hs *HTTPServer) GetDataSources(c *contextmodel.ReqContext) response.Response {
 	query := datasources.GetDataSourcesQuery{OrgID: c.SignedInUser.GetOrgID(), DataSourceLimit: hs.Cfg.DataSourceLimit}
+
+	//BMC Code : Start => MSP Tenant Check only for non admin users
+	if c.SignedInUser.OrgRole != "Admin" && ((len(strings.TrimSpace(c.SignedInUser.SubTenantId)) != 0) || ((len(c.SignedInUser.MspOrgs) > 0) && !c.SignedInUser.IsUnrestrictedUser)) {
+		query = hs.buildMSPDSQuery(c)
+	}
+	//BMC Code : End
 
 	dataSources, err := hs.DataSourcesService.GetDataSources(c.Req.Context(), &query)
 	if err != nil {
@@ -180,6 +187,11 @@ func (hs *HTTPServer) DeleteDataSourceById(c *contextmodel.ReqContext) response.
 		return response.Error(403, "Cannot delete read-only data source", nil)
 	}
 
+	// BMC Code - Disable deletion for non super admin users
+	if (ds.Type == "bmchelix-ade-datasource" || ds.Type == "json-datasource") && !c.IsGrafanaAdmin {
+		return response.Error(403, "You do not have enough permission to delete datasource", nil)
+	}
+
 	cmd := &datasources.DeleteDataSourceCommand{ID: id, OrgID: c.SignedInUser.GetOrgID(), Name: ds.Name}
 
 	err = hs.DataSourcesService.DeleteDataSource(c.Req.Context(), cmd)
@@ -259,10 +271,21 @@ func (hs *HTTPServer) DeleteDataSourceByUID(c *contextmodel.ReqContext) response
 		return response.Error(403, "Cannot delete read-only data source", nil)
 	}
 
+	// BMC Code - Disable deletion for non super admin users
+	if (ds.Type == "bmchelix-ade-datasource" || ds.Type == "json-datasource") && !c.IsGrafanaAdmin {
+		return response.Error(403, "You do not have enough permission to delete datasource", nil)
+	}
+
 	cmd := &datasources.DeleteDataSourceCommand{UID: uid, OrgID: c.SignedInUser.GetOrgID(), Name: ds.Name}
 
 	err = hs.DataSourcesService.DeleteDataSource(c.Req.Context(), cmd)
 	if err != nil {
+		// BMC Code - start
+		kp.DataSourceDeleteEvent.Send(kp.EventOpt{Ctx: c, Err: err,
+			ObjectDetails:    "Datasource of type: " + ds.Type,
+			OperationSubType: "Failed to delete datasource: " + err.Error(),
+		})
+		// BMC Code - end
 		if errors.As(err, &secretsPluginError) {
 			return response.Error(500, "Failed to delete datasource: "+err.Error(), err)
 		}
@@ -271,6 +294,12 @@ func (hs *HTTPServer) DeleteDataSourceByUID(c *contextmodel.ReqContext) response
 
 	hs.Live.HandleDatasourceDelete(c.SignedInUser.GetOrgID(), ds.UID)
 
+	//BMC Code - start
+	kp.DataSourceDeleteEvent.Send(kp.EventOpt{Ctx: c, Prev: nil, New: nil,
+		ObjectDetails:    "Datasource of type: " + ds.Type,
+		OperationSubType: "Datasource deleted successfully",
+	})
+	//BMC Code - end
 	return response.JSON(http.StatusOK, util.DynMap{
 		"message": "Data source deleted",
 		"id":      ds.ID,
@@ -308,6 +337,11 @@ func (hs *HTTPServer) DeleteDataSourceByName(c *contextmodel.ReqContext) respons
 
 	if dataSource.ReadOnly {
 		return response.Error(403, "Cannot delete read-only data source", nil)
+	}
+
+	// BMC Code - Disable deletion for non super admin users
+	if (dataSource.Type == "bmchelix-ade-datasource" || dataSource.Type == "json-datasource") && !c.IsGrafanaAdmin {
+		return response.Error(403, "You do not have enough permission to delete datasource", nil)
 	}
 
 	cmd := &datasources.DeleteDataSourceCommand{Name: name, OrgID: c.SignedInUser.GetOrgID()}
@@ -463,6 +497,13 @@ func (hs *HTTPServer) AddDataSource(c *contextmodel.ReqContext) response.Respons
 
 	dataSource, err := hs.DataSourcesService.AddDataSource(c.Req.Context(), &cmd)
 	if err != nil {
+		//BMC Code - start
+		kp.DataSourceAddEvent.Send(kp.EventOpt{Ctx: c, Err: err,
+			ObjectDetails:    "Datasource of type: " + cmd.Type,
+			OperationSubType: "Failed to add datasource: " + err.Error(),
+		})
+		//BMC Code - end
+
 		if errors.Is(err, datasources.ErrDataSourceNameExists) || errors.Is(err, datasources.ErrDataSourceUidExists) {
 			return response.Error(http.StatusConflict, err.Error(), err)
 		}
@@ -479,6 +520,15 @@ func (hs *HTTPServer) AddDataSource(c *contextmodel.ReqContext) response.Respons
 	hs.accesscontrolService.ClearUserPermissionCache(c.SignedInUser)
 
 	ds := hs.convertModelToDtos(c.Req.Context(), dataSource)
+
+	//BMC Code - start
+	newValue := ds
+	newValue.SecureJsonFields = nil
+	kp.DataSourceAddEvent.Send(kp.EventOpt{Ctx: c, Prev: nil, New: newValue,
+		ObjectDetails:    "Datasource of type: " + cmd.Type,
+		OperationSubType: "Datasource added successfully",
+	})
+	//BMC Code - end
 	return response.JSON(http.StatusOK, util.DynMap{
 		"message":    "Datasource added",
 		"id":         dataSource.ID,
@@ -519,12 +569,22 @@ func (hs *HTTPServer) UpdateDataSourceByID(c *contextmodel.ReqContext) response.
 	if cmd.ID, err = strconv.ParseInt(web.Params(c.Req)[":id"], 10, 64); err != nil {
 		return response.Error(http.StatusBadRequest, "id is invalid", err)
 	}
+	// BMC code
+	// author ateli - Mitigation for SSRF issue - DRJ71-3206
+	platformURL, _ := cmd.JsonData.Get("platformURL").String()
+	if resp := validateSSRF(cmd.Type, cmd.URL, platformURL, cmd.OrgID); resp != nil {
+		return resp
+	}
+	// End
+
 	if resp := validateURL(cmd.Type, cmd.URL); resp != nil {
 		return resp
 	}
+
 	if err := validateJSONData(c.Req.Context(), cmd.JsonData, hs.Cfg, hs.Features); err != nil {
 		return response.Error(http.StatusBadRequest, "Failed to update datasource", err)
 	}
+
 	ds, err := hs.getRawDataSourceById(c.Req.Context(), cmd.ID, cmd.OrgID)
 	if err != nil {
 		if errors.Is(err, datasources.ErrDataSourceNotFound) {
@@ -533,6 +593,10 @@ func (hs *HTTPServer) UpdateDataSourceByID(c *contextmodel.ReqContext) response.
 		return response.Error(500, "Failed to update datasource", err)
 	}
 
+	// BMC Code - Disable deletion for non super admin users
+	if (ds.Type == "bmchelix-ade-datasource" || ds.Type == "json-datasource") && !c.IsGrafanaAdmin {
+		return response.Error(403, "You do not have enough permission to update datasource", nil)
+	}
 	// check if LBAC rules have been modified
 	hasAccess, errAccess := checkTeamHTTPHeaderPermissions(hs, c, ds, cmd)
 	if !hasAccess {
@@ -565,6 +629,15 @@ func (hs *HTTPServer) UpdateDataSourceByUID(c *contextmodel.ReqContext) response
 	}
 	datasourcesLogger.Debug("Received command to update data source", "url", cmd.URL)
 	cmd.OrgID = c.SignedInUser.GetOrgID()
+
+	// BMC code
+	// author ateli - Mitigation for SSRF issue - DRJ71-3206
+	platformURL, _ := cmd.JsonData.Get("platformURL").String()
+	if resp := validateSSRF(cmd.Type, cmd.URL, platformURL, cmd.OrgID); resp != nil {
+		return resp
+	}
+	// End
+
 	if resp := validateURL(cmd.Type, cmd.URL); resp != nil {
 		return resp
 	}
@@ -612,6 +685,16 @@ func checkTeamHTTPHeaderPermissions(hs *HTTPServer, c *contextmodel.ReqContext, 
 }
 
 func (hs *HTTPServer) updateDataSourceByID(c *contextmodel.ReqContext, ds *datasources.DataSource, cmd datasources.UpdateDataSourceCommand) response.Response {
+	//BMC Code - start
+	preValue, errPreValue := hs.getRawDataSourceByUID(c.Req.Context(), cmd.UID, cmd.OrgID)
+	if errPreValue != nil {
+		datasourcesLogger.Error("Failed to get previous values", errPreValue.Error())
+	}
+	if preValue != nil {
+		preValue.SecureJsonData = nil
+	}
+	//End
+
 	if ds.ReadOnly {
 		return response.Error(403, "Cannot update read-only data source", nil)
 	}
@@ -630,6 +713,12 @@ func (hs *HTTPServer) updateDataSourceByID(c *contextmodel.ReqContext, ds *datas
 			return response.Error(http.StatusInternalServerError, "Failed to update datasource: "+err.Error(), err)
 		}
 
+		//BMC Code - start
+		kp.DataSourceUpdateEvent.Send(kp.EventOpt{Ctx: c, Err: err,
+			ObjectDetails:    "Datasource of type: " + cmd.Type,
+			OperationSubType: "Failed to update datasource: " + err.Error(),
+		})
+		//BMC Code - end
 		return response.ErrOrFallback(http.StatusInternalServerError, "Failed to update datasource", err)
 	}
 
@@ -649,7 +738,19 @@ func (hs *HTTPServer) updateDataSourceByID(c *contextmodel.ReqContext, ds *datas
 	datasourceDTO := hs.convertModelToDtos(c.Req.Context(), dataSource)
 
 	hs.Live.HandleDatasourceUpdate(c.SignedInUser.GetOrgID(), datasourceDTO.UID)
-
+	//BMC Code - start
+	newValue, errNewValue := hs.getRawDataSourceByUID(c.Req.Context(), cmd.UID, cmd.OrgID)
+	if errNewValue != nil {
+		datasourcesLogger.Error("Failed to get updated values", errNewValue.Error())
+	}
+	if newValue != nil {
+		newValue.SecureJsonData = nil
+	}
+	kp.DataSourceUpdateEvent.Send(kp.EventOpt{Ctx: c, Prev: preValue, New: newValue,
+		ObjectDetails:    "Datasource of type: " + cmd.Type,
+		OperationSubType: "Datasource updated successfully",
+	})
+	//BMC Code - end
 	return response.JSON(http.StatusOK, util.DynMap{
 		"message":    "Datasource updated",
 		"id":         cmd.ID,
@@ -968,6 +1069,18 @@ func (hs *HTTPServer) checkDatasourceHealth(c *contextmodel.ReqContext, ds *data
 
 	return response.JSON(http.StatusOK, payload)
 }
+
+// BMC code
+// author ateli - Mitigation for SSRF issue - DRJ71-3206
+func validateSSRF(cmdType string, url string, platformURL string, orgId int64) response.Response {
+	if _, err := datasource.ValidateSSRF(cmdType, url, platformURL, orgId); err != nil {
+		return response.Error(400, err.Error(), err)
+	}
+
+	return nil
+}
+
+// End
 
 // swagger:parameters checkDatasourceHealthByID
 type CheckDatasourceHealthByIDParams struct {
